+++ config/headersToFilter.go
@@ -1,0 +1,25 @@
+package config
+
+var defaultFilterHeaders = []string{
+	"Authorization",
+	"Authorization-Info",
+	"Cookie",
+	"Openai-Organization",
+	"Proxy-Authenticate",
+	"Proxy-Authorization",
+	"Set-Cookie",
+	"WWW-Authenticate",
+	"X-Access-Token",
+	"X-Api-Key",
+	"X-Auth-Password",
+	"X-Auth-Token",
+	"X-Auth-User",
+	"X-CSRF-Token",
+	"X-Forwarded-User",
+	"X-Password",
+	"X-Refresh-Token",
+	"X-User-Email",
+	"X-User-Id",
+	"X-UserName",
+	"X-User-Secret",
+}
